Re-structure IKS01A2 sensor configuration
The sensor board has nothing to do with the TARGET_DISCO_L496AG per se,
you can add that same sensor board to literally almost any STM32 target.
Thus, it needs to have it's own configuration.

Tidying up also the code - we do not need to have #ifdefs everywhere.
We just specify the config in the mbed_app.json and have "dummy"
empty implementations in the header files that do nothign - the compiler
actually optimizes them out completely (even if they are "called" in
main.cpp). This is how it is usually done in the Linux world and
this allows for "neater" code flow as it does not require to any

I.e.

```C
        if (setup_security_object() || setup_server_object()
            || device_object_install(anjay)
            || joystick_object_install(anjay)
            || humidity_object_install(anjay)
            || barometer_object_install(anjay)
            || magnetometer_object_install(anjay)
            || accelerometer_object_install(anjay)

```
This split for the configs wit #ifdefs would mean we would
have code like below:

```C
        if (setup_security_object() || setup_server_object()
            || device_object_install(anjay)
            || joystick_object_install(anjay)
            || humidity_object_install(anjay)
            || barometer_object_install(anjay)
            || magnetometer_object_install(anjay)
            || accelerometer_object_install(anjay)
```

We have multiple places where that happens. We centralize it
all to the header file instead.